Reorg to accomodate change
