+++ agixt/Chain.py
@@ -1,34 +1,30 @@
 import os
 import json
-from Interactions import Interactions
 import argparse
-from Extensions import Extensions
 import logging
 from datetime import datetime
 
 
 def create_command_suggestion_chain(agent_name, command_name, command_args):
     chain = Chain()
+    chains = chain.get_chains()
     chain_name = f"{agent_name} Command Suggestions"
-    try:
-        chain_steps = chain.get_chain(chain_name=chain_name)
-        # Get last step number in chain_steps, it is at chain_steps["steps"][-1]["step"]
-        step = int(chain_steps["steps"][-1]["step"]) + 1
-    except:
+    if chain_name in chains:
+        step = int(chain.get_chain(chain_name=chain_name)["steps"][-1]["step"]) + 1
+    else:
         chain.add_chain(chain_name=chain_name)
         step = 1
-        # Add the step to the chain
-        chain.add_chain_step(
-            chain_name=chain_name,
-            agent_name=agent_name,
-            step_number=step,
-            prompt_type="Command",
-            prompt={
-                "command_name": command_name,
-                **command_args,
-            },
-        )
-    return f"AUTONOMOUS_EXECUTION is set to False. The command has been added to a chain called '{agent_name} Command Suggestions' for you to review and execute manually."
+    chain.add_chain_step(
+        chain_name=chain_name,
+        agent_name=agent_name,
+        step_number=step,
+        prompt_type="Command",
+        prompt={
+            "command_name": command_name,
+            **command_args,
+        },
+    )
+    return f"The command has been added to a chain called '{agent_name} Command Suggestions' for you to review and execute manually."
 
 
 def get_chain_file_path(chain_name):
@@ -256,48 +252,6 @@ class Chain:
             if new_prompt_content == {}:
                 new_prompt_content = prompt_content
         return new_prompt_content
-
-    async def run_chain_step(self, step: dict = {}, chain_name="", user_input=""):
-        logging.info(step)
-        if step:
-            if "prompt_type" in step:
-                agent_name = step["agent_name"]
-                agent = Interactions(agent_name)
-                prompt_type = step["prompt_type"]
-                step_number = step["step"]
-                if "prompt_name" in step["prompt"]:
-                    prompt_name = step["prompt"]["prompt_name"]
-                else:
-                    prompt_name = ""
-                args = self.get_step_content(
-                    chain_name=chain_name,
-                    prompt_content=step["prompt"],
-                    user_input=user_input,
-                    agent_name=agent_name,
-                )
-                if prompt_type == "Command":
-                    return await Extensions(
-                        agent_config=agent.agent.agent_config
-                    ).execute_command(
-                        command_name=args["command_name"],
-                        command_args=args,
-                    )
-                elif prompt_type == "Prompt":
-                    result = await agent.run(
-                        user_input=user_input,
-                        prompt=prompt_name,
-                        chain_name=chain_name,
-                        step_number=step_number,
-                        **args,
-                    )
-                elif prompt_type == "Chain":
-                    result = await self.run_chain(
-                        chain_name=step["prompt"]["chain_name"], user_input=user_input
-                    )
-        if result:
-            return result
-        else:
-            return None
 
 
 if __name__ == "__main__":
